V4L/DVB (6963): tda18271: store IF frequency in a u16 instead of u32
Store IF Frequency in a u16 instead of a u32.  Multiply by 1000 before use.

+++ drivers/media/dvb/frontends/tda18271.h
@@ -26,7 +26,7 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 #include "dvb_frontend.h"
 
 struct tda18271_std_map_item {
-	u32 if_freq;
+	u16 if_freq;
 	u8 std_bits;
 };
